fix: fix prettifier
